dapp/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-ckeditor-5==0.2.12
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-parler==2.3
django-parler-rest==2.2
django-resized==1.0.2
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.4.4
